requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
supabase>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
celery>=5.3,<6.0
redis>=5.0,<6.0
openai>=1.0,<2.0
langchain>=0.1,<0.3
langchain-openai>=0.0,<0.2
python-docx>=1.0,<2.0
pdfplumber>=0.10,<0.11
python-dotenv>=1.0,<2.0
django-cors-headers>=4.3,<5.0
gunicorn>=21.0,<22.0